fix(breadcrumbs): Link transactions in breadcrumbs to performance
Not all users have access to discover, so link transactions in breadcrumbs to
performance where more users have access to it.
