+++ static/app/views/performance/compare/transactionSummary.tsx
@@ -9,8 +9,7 @@ import {t} from 'sentry/locale';
 import space from 'sentry/styles/space';
 import {Organization} from 'sentry/types';
 import {Event} from 'sentry/types/event';
-
-import {getTransactionDetailsUrl} from '../utils';
+import {getTransactionDetailsUrl} from 'sentry/utils/performance/urls';
 
 import {isTransactionEvent} from './utils';
 
@@ -53,7 +52,7 @@ class TransactionSummary extends Component<Props> {
                 <span>{t('ID')}: </span>
                 <StyledLink
                   to={getTransactionDetailsUrl(
-                    organization,
+                    organization.slug,
                     baselineEventSlug.trim(),
                     baselineEvent.title,
                     location.query
@@ -77,7 +76,7 @@ class TransactionSummary extends Component<Props> {
                 <span>{t('ID')}: </span>
                 <StyledLink
                   to={getTransactionDetailsUrl(
-                    organization,
+                    organization.slug,
                     regressionEventSlug.trim(),
                     regressionEvent.title,
                     location.query
